requirements.txt
discord.py~=2.3.2
python-dotenv~=0.19.2